Renamed protected-column to encrypted-column
