+++ stashjs-typeorm/src/index.ts
@@ -1,4 +1,4 @@
-export { ProtectedColumn, Queryable, ConfigurationMetadata } from "./protected-column"
+export { EncryptedColumn, Queryable, ConfigurationMetadata } from "./encrypted-column"
 export { CollectionManager } from "./collection-manager"
 export { IndexingSubscriber } from "./indexing-subscriber"
 export { wrapRepo, WrappedRepository } from "./repo-extension"
